+++ lib/plausible/stats/current_visitors.ex
@@ -2,10 +2,10 @@ defmodule Plausible.Stats.CurrentVisitors do
   use Plausible.ClickhouseRepo
   use Plausible.Stats.SQL.Fragments
 
-  def current_visitors(site) do
+  def current_visitors(site, duration \\ Duration.new(minute: -5)) do
     first_datetime =
       NaiveDateTime.utc_now()
-      |> NaiveDateTime.shift(minute: -5)
+      |> NaiveDateTime.shift(duration)
       |> NaiveDateTime.truncate(:second)
 
     ClickhouseRepo.one(